Moved Shape classes into separate file
The main file is getting to the size where we need to start separating things into logical units/modules. I moved the Shape classes into their own file so that, in the future, when/if we want to make more graphics/shape related additions, it can remain isolated to this file (i.e. it won't impact the main file too heavily).

I also changed the name of the main game loop file to be main.py.
